Fixed up LD_LIBRARY_PATH setting.

+++ lib/language_pack/ruby.rb
@@ -489,14 +489,12 @@ WARNING
   def install_oci8_binaries
     
     `mkdir -p #{ORACLE_INSTANT_CLIENT_DIR}` unless Dir.exists?(ORACLE_INSTANT_CLIENT_DIR)
-    `mkdir -p #{ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH}` unless Dir.exists?(ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH)
-    puts "ORACLE_INSTANT_CLIENT_DIR is #{ORACLE_INSTANT_CLIENT_DIR}"
-    puts "ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH is #{ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH}"
+    raise "Dead code" unless Dir.exists?(ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH)
     
     result = `curl #{ORACLE_INSTANT_CLIENT_TGZ_URL} -s -o - | tar -xz -C #{ORACLE_INSTANT_CLIENT_DIR} -f - `
     if $?.success?
       puts "Setting OCI8 environment variables"
-      ENV["LD_LIBRARY_PATH"]="#{ORACLE_INSTANT_CLIENT_DIR_ABSOLUTE_PATH}:$LD_LIBRARY_PATH"
+      ENV["LD_LIBRARY_PATH"]="#{ORACLE_INSTANT_CLIENT_DIR}:#{ENV['LD_LIBRARY_PATH']}"
       ENV["NLS_LANG"]='AMERICAN_AMERICA.UTF8'
       puts "Done installing OCI8"
     else
